feat: organize all component directories consistently
