+++ src/hoc/index.js
@@ -3,6 +3,8 @@
  * Exports all hoc components.
  */
 
-import ThemeProvider from './ThemeProvider';
+import ThemeProvider from './ThemeProvider/ThemeProvider';
+import PrivateRoute from './PrivateRoute/PrivateRoute.container';
+import PublicRoute from './PublicRoute/PublicRoute.container';
 
-export { ThemeProvider };
+export { ThemeProvider, PrivateRoute, PublicRoute };
