+++ src/layouts/ThinLayout/ThinLayout.style.js
