+++ src/layouts/index.js
@@ -3,6 +3,8 @@
  * Exports all layout components.
  */
 
-import ThinLayout from './ThinLayout';
+import ThinLayout from './ThinLayout/ThinLayout';
+import Header from './Header/Header';
+import Footer from './Footer/Footer';
 
-export { ThinLayout };
+export { ThinLayout, Header, Footer };
