+++ src/pages/index.js
@@ -5,5 +5,6 @@
 
 import Home from './Home/Home';
 import Login from './Login/Login';
+import Dashboard from './Dashboard/Dashboard';
 
-export { Home, Login };
+export { Home, Login, Dashboard };
